requirements.txt
pillow==11.3.0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pyopencl==2025.1
superqt==0.7.6
typing_extensions==4.14.1
